Add more scripts. Update install.sh

+++ install.sh
@@ -1,6 +1,19 @@
 #!/bin/bash
 
+die() { echo "$@"; exit 1; }
+
+U="$1"
+id -u "$U" || die "Please supply your main non-root user on the command line."
+[ -f ./install.sh ] || die "Please run the script from its directory."
+
 set -uxeo pipefail
+
+### Set console font size
+
+perl -npi -E '/FONTSIZE=/ and $_=qq{FONTSIZE="8x14zz"\n}' /etc/default/console-setup
+setupcon
+
+### Install packages
 
 apt update -y
 DEBIAN_FRONTEND=noninteractive apt install -y \
@@ -14,6 +27,19 @@ DEBIAN_FRONTEND=noninteractive apt install -y \
     ffmpeg \
     youtube-dl \
     #
+
+### Seed user's home
+
+USERHOME="$(bash -c "echo ~$U")"
+-d "$USERHOME" || exit 1
+[[ "$USERHOME" != "$HOME" ]] && cp -a home/.[a-z]* "$USERHOME"
+
+### Mount points
+
+mkdir -p /mnt/data
+mkdir -p /mnt/zoom
+
+### Set up Samba
 
 cat >> /etc/samba/smb.conf << '_E'
 
@@ -45,6 +71,8 @@ cat >> /etc/samba/smb.conf << '_E'
 _E
 service smbd restart
 
+### Add an alias for usb drives.
+
 cat > /etc/udev/rules.d/80-usbdev.rules << '_E'
 # /etc/udev/rules.d/80-usbdev.rules
 # https://community.openhab.org/t/how-to-make-symlinks-for-usb-ports-in-linux-extra-java-opts/89615
@@ -69,10 +97,41 @@ udevadm control -R
 # /usr/lib/udev/rules.d/...
 # udisksctl mount -b /dev/sdc1
 
-# /etc/crontab:
-# */5 *   * * *   ershov  /usr/bin/flock -n /home/ershov/nas/phcp-cron-nas.sh /home/ershov/nas/phcp-cron-nas.sh
+### Allow user to mount the flash drive:
 
-perl -npi -E '/FONTSIZE=/ and $_=qq{FONTSIZE="8x14zz"\n}' /etc/default/console-setup
-setupcon
+#mkdir -p /mnt/zoom
+#cat >> /etc/fstab << '_E'
+#/dev/sdd1 /mnt/zoom/ vfat ro,user,noauto 0 1
+#_E
 
+### Copy scripts
+
+cp -a bin/* /usr/local/bin/
+
+### Set up touch screen (if present)
+
+if lsusb |& fgrep -q USB3IIC_CTP_CONTROL; then
+    cp sbin/touchscreen_track_USB2IIC_CTP_CONTROL.sh /usr/local/bin/
+    chown 0:0 sbin/touchscreen_track_USB2IIC_CTP_CONTROL.sh
+    cat > /etc/sudoers.d/touch << "_E"
+# Allow do dump touch events on the console
+
+User_Alias TOUCH_USERS = $U
+
+TOUCH_USERS ALL = NOPASSWD: /usr/local/bin/touchscreen_track_USB2IIC_CTP_CONTROL.sh
+
+_E
+fi
+
+
+### Set up crontab actions
+
+cat > /etc/cron.d/clean_macos_files << "_E"
+SHELL=/bin/sh
+*/5 *   * * *   root    test -d /mnt/data/incoming && find /mnt/data/incoming -name ._\* -delete
+_E
+cat > /etc/cron.d/phcp << "_E"
+SHELL=/bin/sh
+*/5 *   * * *   $U  /usr/bin/flock -n /usr/local/bin/phcp-cron-nas.sh /usr/local/bin/phcp-cron-nas.sh
+_E
